kernel: move more error related code to error.c
