+++ src/error.h
@@ -146,11 +146,11 @@ extern void ErrorReturnVoid (
             Int                 arg2,
             const Char *        msg2 );
 
+
 /****************************************************************************
 **
 *F * * * * * * * * * * * * * * initialize module * * * * * * * * * * * * * *
 */
-
 
 /****************************************************************************
 **
